Diabetic-retinopathy-detection-main/requirements.txt
torch
torchvision
gradio
pillow
matplotlib
numpy
pandas
scikit-learn
